+++ CTPCompareTool/include/MdSpiEx.h
@@ -1,21 +1,43 @@
 #pragma once
 
+#include "ThostFtdcMdApi.h"
+
 #include "TickMatcher.h"
 #include "TickConverter.h"
 
-//======================================================
-// MdSpiEx
-//======================================================
-
-class MdSpiEx
+class MdSpiEx : public CThostFtdcMdSpi
 {
 public:
 
-    MdSpiEx(TickMatcher& matcher);
+    MdSpiEx(
+        LineType line,
+        TickMatcher& matcher);
 
-    void OnRtnDepthMarketData(const Tick* tick);
+public:
+
+    virtual void OnFrontConnected() override;
+
+    virtual void OnFrontDisconnected(int nReason) override;
+
+    virtual void OnRspUserLogin(
+        CThostFtdcRspUserLoginField* pRspUserLogin,
+        CThostFtdcRspInfoField* pRspInfo,
+        int nRequestID,
+        bool bIsLast) override;
+
+    virtual void OnRtnDepthMarketData(
+        CThostFtdcDepthMarketDataField* pData) override;
+
+public:
+    void SetApi(CThostFtdcMdApi* api);
 
 private:
 
+    LineType      m_line;
+
     TickMatcher& m_matcher;
+
+    TickConverter m_converter;
+
+    CThostFtdcMdApi* m_api = nullptr;
 };